Sorry didn't find a way to comment.
This is great and runs like crazy. I installed it with Rufus as Windows to Go on an external USB SSD. But now I have executed the two included batch files, and the WiFi of the laptop no longer works. Is there a way to restore WiFi functionality without reinstalling Windows?  Thank you!